materials list on diagram page

+++ Domekit/Views/GeodesicView.h
@@ -32,7 +32,7 @@ typedef enum{
 
 @property BOOL showMeridians;
 
-@property float sliceAmount;
+@property unsigned int sliceAmount;
 
 -(void) drawTriangles;
